Single responsibility added and open close start
Finished doing the example of Single responsibility.
Started to work on Open close Solid.

Added Fireball to mario.
