+++ Assets/Scripts/SC_FireBall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8566d08b77df014468cfac0571ce7a64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
