+++ Assets/Scripts/SC_Movement.cs
@@ -1,0 +1,28 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SC_Movement : MonoBehaviour
+{
+    public float speed = 5;
+    private float direction;
+    private Rigidbody2D rigid;
+
+    void Awake()
+    {
+        rigid = GetComponent<Rigidbody2D>();
+    }
+    void FixedUpdate()
+    {
+        direction = Input.GetAxis("Horizontal");
+        if (direction != 0 && rigid != null)
+        {
+            rigid.velocity = new Vector2(direction * speed, rigid.velocity.y);
+
+            if (direction > 0)
+                transform.localScale = new Vector3(1, 1, 1);
+            else transform.localScale = new Vector3(-1, 1, 1);
+        }
+    }
+
+}
