+++ Assets/Scripts/SC_Movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03112e3b90009b84d8171221a05f62a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
