+++ Assets/Scripts/SC_ResetPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dba2a4ee26bc77845b85ad4559b0847d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
