+++ Assets/Scripts/SC_Weapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8980178a5116a064399d39008302144c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
